Full transition to Firebase
This is a doozy of a change, but it builds successfully. This may justify a hard fork, since GCM is no longer relevant.  Unsure if the module as built will actually work, and what the implications are for backwards compatibility.

Libs updated, version bumped, and build files removed since they are no longer needed after Ti SDK 7+.