Wait for token before sending success callback
IIDService now broadcasts an Intent which GCMModule listens for. If during registration the token is not yet available, a flag is set, and once the token is refreshed the success callback is called. The way we aren't sending an error immediately, but there is no guarantee that success or error will be called.